1.增加数据库配置database.properties 2.增加主机配置相关接口

+++ src/main/java/tpf/dao/DatabasePOMapper.java
@@ -3,13 +3,13 @@ package tpf.dao;
 import tpf.pojo.DatabasePO;
 
 public interface DatabasePOMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(DatabasePO record);
 
     int insertSelective(DatabasePO record);
 
-    DatabasePO selectByPrimaryKey(Long id);
+    DatabasePO selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(DatabasePO record);
